Fix broken symlinks when old_slug contains dirs
Without this patch, the symlink is created at the old_slug location
with the target being the new slug's relative path from the jekyll
root directory. Of course, that only works if the old_slug was in the
jekyll root directory.

This fix finds the relative path from the old_slug to the new slug,
and specifies that path as the link target.